More checks in the chicken feeder. Prevents chicken feeder to go into the negative numbers in inventory
Nest Pen now drops items and chicken if broken
Nest Pen now stores the egg it drops and makes copies. This is to prevent inventories filling with same egg, because nbt data is a bit different. This is a temp fix, and will still cause dup when reloading chunks or worlds. but will minimize it.